fix(usage): split real-time gauges onto stats-events
Hourly full-count jobs and 60-second gauges cannot share one list or the
gauges starve. Rename the existing consumer to stats-calculations and add
stats-events. Keep the StatsResources type and inject name so Cloud can
extend this release without a Composer pin bump.
