+++ .env
@@ -153,6 +153,8 @@ _APP_CONNECTIONS_DB_USAGE=http://appwrite:appwrite@clickhouse:8123/appwrite
 _APP_POOL_SIZE_USAGE=2
 _APP_MAINTENANCE_RETENTION_USAGE_TTL=180
 _APP_STATS_USAGE_QUEUE_NAME=v1-stats-usage
+_APP_STATS_CALCULATIONS_QUEUE_NAME=v1-stats-calculations
+_APP_STATS_EVENTS_QUEUE_NAME=v1-stats-events
 _APP_STATS_RESOURCES_QUEUE_NAME=v1-stats-resources
 _APP_STATS_RESOURCES_INTERVAL=3600
 _APP_INTERVAL_CLEANUP_STALE_EXECUTIONS=300
