+++ bin/worker-stats-events
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+exec php /usr/src/code/app/worker.php stats-events "$@"
