+++ src/Appwrite/Platform/Modules/Usage/Services/Tasks.php
@@ -2,7 +2,7 @@
 
 namespace Appwrite\Platform\Modules\Usage\Services;
 
-use Appwrite\Platform\Tasks\StatsResources;
+use Appwrite\Platform\Tasks\StatsCalculations;
 use Appwrite\Platform\Tasks\UsageSetup;
 use Utopia\Platform\Service;
 
@@ -11,7 +11,7 @@ class Tasks extends Service
     public function __construct()
     {
         $this->type = Service::TYPE_TASK;
-        $this->addAction(StatsResources::getName(), new StatsResources());
+        $this->addAction(StatsCalculations::getName(), new StatsCalculations());
         $this->addAction(UsageSetup::getName(), new UsageSetup());
     }
 }
