+++ src/Appwrite/Platform/Modules/Usage/Services/Workers.php
@@ -2,7 +2,8 @@
 
 namespace Appwrite\Platform\Modules\Usage\Services;
 
-use Appwrite\Platform\Workers\StatsResources;
+use Appwrite\Platform\Workers\StatsCalculations;
+use Appwrite\Platform\Workers\StatsEvents;
 use Appwrite\Platform\Workers\StatsUsage;
 use Utopia\Platform\Service;
 
@@ -12,6 +13,7 @@ class Workers extends Service
     {
         $this->type = Service::TYPE_WORKER;
         $this->addAction(StatsUsage::getName(), new StatsUsage());
-        $this->addAction(StatsResources::getName(), new StatsResources());
+        $this->addAction(StatsCalculations::getName(), new StatsCalculations());
+        $this->addAction(StatsEvents::getName(), new StatsEvents());
     }
 }
