+++ src/Appwrite/Platform/Tasks/Schedule.php
@@ -95,7 +95,7 @@ class Schedule extends Action
         $this->loop(fn () => (new ScheduleMessages())->action($publisherForMessaging, $getIsResourceBlocked, $dbForPlatform, $getProjectDB, $telemetry, $locks));
 
         if ($publisherForStatsResources !== null && $usageConnection !== null) {
-            $this->loop(fn () => (new StatsResources())->action($dbForPlatform, $publisherForStatsResources, $usageConnection));
+            $this->loop(fn () => (new StatsCalculations())->action($dbForPlatform, $publisherForStatsResources, $usageConnection));
         }
 
         while (true) {
